some refactoring/reformatting
