+++ plugins/com.google.cloud.tools.eclipse.appengine.libraries/src/com/google/cloud/tools/eclipse/appengine/libraries/LibraryClasspathContainer.java
@@ -27,12 +27,14 @@ public class LibraryClasspathContainer implements IClasspathContainer {
   private final String description;
   private final IClasspathEntry[] classpathEntries;
 
-  public LibraryClasspathContainer(IPath path, String description, IClasspathEntry[] classpathEntries) {
+  public LibraryClasspathContainer(IPath path, String description,
+      IClasspathEntry[] classpathEntries) {
     Preconditions.checkNotNull(path, "path is null");
     Preconditions.checkNotNull(description, "description is null");
     Preconditions.checkArgument(!description.isEmpty(), "description is empty");
     Preconditions.checkNotNull(classpathEntries, "classpathEntries is null");
-    containerPath = path;
+    
+    this.containerPath = path;
     this.description = description;
     this.classpathEntries = classpathEntries;
   }
@@ -40,6 +42,7 @@ public class LibraryClasspathContainer implements IClasspathContainer {
   /**
    * Creates a new {@link LibraryClasspathContainer} with the same path and description,
    * but with the <code>classpathEntries</code>.
+   * 
    * @param classpathEntries the classpath entries of the new container
    */
   public LibraryClasspathContainer copyWithNewEntries(IClasspathEntry[] classpathEntries) {
